Changed dtype to float32

+++ src/data_preprocessing/baseline_lstm_preprocessing.py
@@ -11,11 +11,11 @@ def get_train_test(df, window=24, train_size=0.8):
     for j in range(len(df) - window):
         X.append(df["price actual"][j: j + window])
         y.append(df["price actual"].iloc[j + window])
-    X = np.array(X)
-    y = np.array(y)
+    X = np.array(X).astype(np.float32)
+    y = np.array(y).astype(np.float32)
     x_train = X[:int(len(X) * train_size)]
     y_train = y[:int(len(y) * train_size)]
-    valid_size = int(len(X) - len(x_train) / 2)
+    valid_size = int((len(X) - len(x_train)) / 2)
     x_valid = X[int(len(X) * train_size): int(len(X) * train_size) + valid_size]
     y_valid = y[int(len(y) * train_size): int(len(y) * train_size) + valid_size]
     x_test = X[int(len(X) * train_size) + valid_size:]
